Backend/requirements.txt
python-docx
PyPDF2
reportlab
typing
PyMuPDF
langchain-google-genai
python-dotenv
gliner
flask